fixing for devpac / Darwin setup

+++ buildrepl.sh
@@ -1,23 +1,42 @@
-# assumes that "vasm" is in your PATH
-vasm -no-opt -Fbin -o replayer/MYM_REPL.BIN replay.s/stubrepl.s
+# detect if "vasm" is in your PATH
+if command -v vasm >&2; then
+    echo vasm detected
 
-# Probably a bug in vasm? The output file "MYM_REPL.PRG" is created instead of "MYM_REPL.BIN"
-mv replayer/MYM_REPL.PRG replayer/MYM_REPL.BIN
+    vasm -no-opt -Fbin -o replayer/MYM_REPL.BIN replay.s/stubrepl.s
+    # Probably a bug in vasm? The output file "MYM_REPL.PRG" is created instead of "MYM_REPL.BIN"
+    mv replayer/MYM_REPL.PRG replayer/MYM_REPL.BIN
 
-cd replayer
-vasm -no-opt -Fbin -o MULTSNDH.SND MULTSNDH.S
-cd ..
+    cd replayer
+    vasm -no-opt -Fbin -o MULTSNDH.SND MULTSNDH.S
+    cd ..
+    exit 0
+fi
 
-exit 0
+# detect if crosstos/dist/devpac is in your PATH    
+if command -v m68k-atari-tos-devpac-gen >&2; then
+    echo crosstos devpac detected
 
-# Version that builds the files using crosstos devpac
-# assumes that crosstos/dist/devpac is in your PATH
+    unameOut="$(uname -s)"
 
-m68k-atari-tos-devpac-gen ./replay.s/stubrepl.s
-tail -c +29 ./replay.s/stubrepl.PRG | head -c -4 > ./replayer/MYM_REPL.BIN
-rm ./replay.s/stubrepl.PRG
+    m68k-atari-tos-devpac-gen ./replay.s/stubrepl.s
+    tail -c +29 ./replay.s/stubrepl.PRG > ./replay.s/stubrepl.BIN
+    if [ "$unameOut" = "Darwin" ]; then
+        newfsize=$(expr $(stat -f '%z' ./replay.s/stubrepl.BIN) - 4)    
+        dd if=./replay.s/stubrepl.BIN of=./replay.s/MYM_REPL.BIN bs=1 count=$newfsize
+    else
+        head -c -4 ./replay.s/stubrepl.BIN > ./replay.s/MYM_REPL.BIN
+    fi
+    rm ./replay.s/stubrepl.BIN ./replay.s/stubrepl.PRG
+    mv ./replay.s/MYM_REPL.BIN ./replayer/
 
-cd ./replayer/
-m68k-atari-tos-devpac-gen ./MULTSNDH.S
-cat MULTSNDH.PRG | tail -c +29 > MULTSNDH.SND
-rm MULTSNDH.PRG
+    cd ./replayer/
+    rm ./MULTSNDH.SND
+    m68k-atari-tos-devpac-gen ./MULTSNDH.S
+    cat MULTSNDH.PRG | tail -c +29 > MULTSNDH.SND
+    rm MULTSNDH.PRG
+    cd ..
+    exit 0
+fi
+
+echo no supported assembler on path
+exit 1
